requirements.txt
requests>=2.31.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
elevenlabs>=0.2.0
pydub>=0.25.1